Rename Cardano.CLI.Shelley.Output to Cardano.CLI.Types.Output
